Fix on off status loop (#87)
* Amend {On_Off}_Status definition to fix inconsistencies

As pointed out in #86, the existing definition of these classes has
On_Status, Off_Status and On_Off_Status as equivalent. This commit fixes
that so that On_Status and Off_Status are composed to form
On_Off_Status. This remedies the inconsistency in the class lattice and
passess all of the tests

* regenerating Brick