Change Brick ontology versioning to be major.minor (ignore patch)
Encoding the version of the ontology into the URI is important for
indicating to users which versions of the ontology are "backwards
compatible", but we do not want users to have to change their URIs
often. I think that encoding the MAJOR and MINOR revision numbers into
the URI (ignoring that PATCH revision) is a nice intermediate point.

This commit encodes the MAJOR, MINOR and PATCH version numbers in
`bricksrc/namespaces.py`, and exposes the MAJOR_MINOR and FULL_VERSION
constants that provide the full semantic version.