Working on SHACL shapes to detect violations of disjoint classes
One of the issues with validating the use of Brick properties is that
the same RDFS inferencing that we use to determine the types of the
entities in the graph also infers extra classes from the use of the
properties. This means that we can't necessarily use a SHACL shape to
make sure that the class of the object of a `hasLocation` relationship
is actually a Location, because the use of the `hasLocation`
relationship (and specifically, the `rdfs:range` property) infers the
object of the property as a Location, even if it is already an instance
of another root class like Equipment.

What these new shapes do is generate a SHACL shape for each pair of root
classes to detect any violations of the disjoint properties.

There are also some small code style changes included in this commit